+++ soohyun/first/RuiBao.kt
@@ -1,0 +1,5 @@
+package kotlinstudy
+
+fun main(){
+    print("대충 루이바오라는 뜻")
+}
